style: Move the code related to layers to a separate file
